Move HighAccuracyTLEPredictor to predictors package
