+++ orbit_predictor/predictors/__init__.py
@@ -1,3 +1,4 @@
 from orbit_predictor.predictors.base import Position, PredictedPass  # noqa
 from orbit_predictor.exceptions import NotReachable  # noqa
 from orbit_predictor.predictors.tle import TLEPredictor  # noqa
+from orbit_predictor.predictors.accurate import HighAccuracyTLEPredictor  # noqa
