+++ tests/test_accurate_predictor.py
@@ -259,9 +259,9 @@ class AccuratePredictorCalculationErrorTests(TestCase):
         # Predictor
         self.predictor = HighAccuracyTLEPredictor(BUGSAT_SATE_ID, self.db)
         self.is_ascending_mock = self._patch(
-            'orbit_predictor.accuratepredictor.LocationPredictor.is_ascending')
+            'orbit_predictor.predictors.accurate.LocationPredictor.is_ascending')
         self.start = datetime(2017, 3, 6, 7, 51)
-        logassert.setup(self,  'orbit_predictor.accuratepredictor')
+        logassert.setup(self,  'orbit_predictor.predictors.accurate')
 
     def _patch(self, *args,  **kwargs):
         patcher = mock.patch(*args, **kwargs)
